+++ src/main/java/com/springboot/couchbase/springbootrealworld/domain/article/service/FavoriteServiceImpl.java
@@ -1,17 +1,22 @@
 package com.springboot.couchbase.springbootrealworld.domain.article.service;
 
+import com.springboot.couchbase.springbootrealworld.domain.article.dto.ArticleDto;
 import com.springboot.couchbase.springbootrealworld.domain.article.dto.FavoriteDto;
+import com.springboot.couchbase.springbootrealworld.domain.article.entity.ArticleDocument;
 import com.springboot.couchbase.springbootrealworld.domain.article.entity.FavoriteDocument;
 import com.springboot.couchbase.springbootrealworld.domain.article.repository.ArticleRepository;
 import com.springboot.couchbase.springbootrealworld.domain.article.repository.FavoriteRepository;
 import com.springboot.couchbase.springbootrealworld.domain.profile.dto.ProfileDto;
 import com.springboot.couchbase.springbootrealworld.domain.profile.service.ProfileService;
 import com.springboot.couchbase.springbootrealworld.domain.user.repository.UserRepository;
+import com.springboot.couchbase.springbootrealworld.exception.AppException;
+import com.springboot.couchbase.springbootrealworld.exception.Error;
 import com.springboot.couchbase.springbootrealworld.security.AuthUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -20,6 +25,10 @@ import java.util.stream.Collectors;
 public class FavoriteServiceImpl implements FavoriteService {
     @Autowired
     private ArticleRepository articleRepository;
+
+    @Autowired
+    private final ArticleService articleService;
+
     @Autowired
     private FavoriteRepository favoriteRepository;
     @Autowired
@@ -38,12 +47,13 @@ public class FavoriteServiceImpl implements FavoriteService {
 
     @Transactional
     @Override
-    public void delete(String slug, AuthUserDetails authUserDetails) {
-
-        String articleId = articleRepository.findBySlug(slug).getId();
+    public ArticleDto delete(String slug, AuthUserDetails authUserDetails) {
+        ArticleDocument article = articleRepository.findBySlug(slug);
+        if (article == null) throw new AppException(Error.ARTICLE_NOT_FOUND);
+        String articleId = article.getId();
         List<FavoriteDocument> favoriteEntities = favoriteRepository.findByArticleId(articleId);
         favoriteRepository.deleteAll(favoriteEntities);
-
+        return articleService.getArticle(slug, authUserDetails);
     }
 
     private FavoriteDto convertToDTO(AuthUserDetails authUserDetails, FavoriteDocument favoriteDocument) {
